gui-sdl: write_text: use beginning and end of too long strings
previously it just wrote outside the screen buffer when encountering
a long filename.

now the first and last 20 bytes are taken and the middle part replaced
with "...". this allows to see the file extension and disk number,
which is usually at the end, as well as the first part of the game
name.