make uaecptr 32bit even on 64bit archs
this type is only used to point into the amiga memory.
see https://github.com/tonioni/WinUAE/pull/5 for more information.

additionally we make the 64bit types always typedef'd to long long,
where possible (lack of long long is only possible on broken pre-
C99 compilers), this has the advantage that we can always use the "ll"
format specifier when dealing with those types, which would result in
warnings if the implementation typedefs int64_t from stdint to long
on 64bit archs (which is the case for typical linux gcc setups).